+++ src/interpol/basis.c
@@ -408,7 +408,7 @@ static int compute_tensor_from_sum_degree_rec(PnlMatInt *T, int degree, int nb_v
 
 static PnlMatInt *compute_tensor_from_sum_degree(int degree, int nb_variates)
 {
-  int nb_elements = pnl_round(pnl_sf_choose(nb_variates + degree, degree));
+  int nb_elements = (int) pnl_round(pnl_sf_choose(nb_variates + degree, degree));
   PnlMatInt *T = pnl_mat_int_create(nb_elements, nb_variates);
   compute_tensor_from_sum_degree_rec(T, degree, nb_variates);
   return T;
@@ -435,7 +435,7 @@ static PnlMatInt *compute_tensor_from_hyperbolic_degree(double degree, double q,
   int i, i_sparse;
   double degree_q;
   PnlMatInt *T;
-  T = compute_tensor_from_sum_degree(ceil(degree), n);
+  T = compute_tensor_from_sum_degree((int)ceil(degree), n);
   degree_q = pow(degree, q);
   for (i = 0, i_sparse = 0 ; i < T->m ; i++)
     {
